doc-qwen3.5-27b/backend/requirements.txt
# ── Core backend ──────────────────────────────────────────
fastapi
uvicorn[standard]
python-multipart

# ── PDF processing ─────────────────────────────────────────
pymupdf
pillow

# ── HF inference client ────────────────────────────────────
openai>=1.0.0
python-dotenv
